+++ src/templates/stacked_skeleton/app/router/router.ts
@@ -1,5 +1,6 @@
 import * as _ from "lodash";
-import { Base } from "../../../base";
+import { Utils } from "../../../../utils/utils";
+import { Base } from "../../base";
 
 export class Router extends Base {
   private _dartString: string;
@@ -7,20 +8,21 @@ export class Router extends Base {
   constructor(fileName: string, projectName: string, suffix?: string) {
     super(fileName, suffix);
 
-    this._dartString = `import 'package:auto_route/auto_route_annotations.dart';
+    this._dartString = `import 'package:auto_route/auto_route.dart';
 
-import 'package:${projectName}/src/ui/views/startup/startup_view.dart';
-import 'package:${projectName}/src/ui/views/home/home_view.dart';
+import 'package:${projectName}/ui/views/home/home_view.dart';
+import 'package:${projectName}/ui/views/startup/startup_view.dart';
 
 export './router.gr.dart';
 
 @AdaptiveAutoRouter(
+  replaceInRouteName: "View,Route",
   routes: <AutoRoute>[
     AdaptiveRoute(page: StartupView, initial: true),
     AdaptiveRoute(page: HomeView),
   ],
 )
-class $AppRouter {}`;
+class $${Utils.convertToPascalCase(projectName)}Router {}`;
   }
 
   get dartString(): string {