+++ src/templates/stacked_skeleton/app/utils/logger.ts
@@ -26,10 +26,10 @@ class SimpleLogPrinter extends LogPrinter {
 
   @override
   List<String> log(LogEvent event) {
-    var color = PrettyPrinter.levelColors[event.level];
-    var emoji = PrettyPrinter.levelEmojis[event.level];
+    AnsiColor? color = PrettyPrinter.levelColors[event.level];
+    String? emoji = PrettyPrinter.levelEmojis[event.level];
     print(
-      color(
+      color!(
         '\$emoji \$className - \${event.message}',
       ),
     );