+++ src/templates/stacked_skeleton/ui/views/startup_view.ts
@@ -9,9 +9,8 @@ export class StartupView extends Base {
 
     this._dartString = `import 'package:flutter/material.dart';
 import 'package:stacked/stacked.dart';
-import 'package:${YamlHelper.getProjectName()}/src/ui/widgets/dumb/skeleton.dart';
-import 'package:${YamlHelper.getProjectName()}/src/ui/global/ui_helpers.dart';
-import 'package:${YamlHelper.getProjectName()}/src/ui/views/startup/startup_view_model.dart';
+
+import './startup_view_model.dart';
 
 class StartupView extends StatefulWidget {
   @override
@@ -23,14 +22,15 @@ class _StartupViewState extends State<StartupView> {
   Widget build(BuildContext context) {
     return ViewModelBuilder<StartupViewModel>.reactive(
       viewModelBuilder: () => StartupViewModel(),
-      onModelReady: (StartupViewModel model) => model.handleStartup(),
+      onModelReady: (StartupViewModel model) async {
+        await model.init();
+      },
       builder: (
         BuildContext context,
         StartupViewModel model,
-        Widget child,
+        Widget? child,
       ) {
-        return Skeleton(
-          isBusy: model.isBusy,
+        return Scaffold(
           body: Center(
             child: Column(
               mainAxisAlignment: MainAxisAlignment.center,
@@ -38,7 +38,7 @@ class _StartupViewState extends State<StartupView> {
               children: <Widget>[
                 // TODO(yazeed): Put Your Logo Here :)
 
-                verticalSpaceMedium(context),
+                const SizedBox(height: 15),
                 CircularProgressIndicator(),
               ],
             ),