+++ src/templates/stacked_skeleton/ui/views/startup_view_model.ts
@@ -7,29 +7,25 @@ export class StartupViewModel extends Base {
   constructor(fileName: string, suffix: string) {
     super(fileName, suffix);
 
-    this._dartString = `import 'dart:async';
-
-import 'package:stacked_services/stacked_services.dart';
-
-import 'package:${YamlHelper.getProjectName()}/src/app/generated/locator/locator.dart';
-import 'package:${YamlHelper.getProjectName()}/src/app/generated/router/router.gr.dart';
-import 'package:${YamlHelper.getProjectName()}/src/ui/global/custom_base_view_model.dart';
+    this._dartString = `import 'package:flutter/material.dart';
+import 'package:${YamlHelper.getProjectName()}/app/core/custom_base_view_model.dart';
+import 'package:${YamlHelper.getProjectName()}/app/locator/locator.dart';
+import 'package:${YamlHelper.getProjectName()}/app/router/router.dart';
+import 'package:${YamlHelper.getProjectName()}/app/services/router_service.dart';
 
 class StartupViewModel extends CustomBaseViewModel {
-  final NavigationService _navigationService = locator<NavigationService>();
-  Future handleStartup() async {
-    /// Do Some Logic Here
-    /// The timer is a placeholder, but the view needs to be viewed at least for a second!
-    Timer(
-      Duration(
-        seconds: 1,
-      ),
-      () async => await navigateToHomeView(),
-    );
+  final RouterService _routerService = locator<RouterService>();
+
+  Future<void> init() async {
+    WidgetsBinding.instance!.addPostFrameCallback((Duration duration) async {
+      await navigateToHomeView();
+    });
   }
 
   Future navigateToHomeView() async {
-    await _navigationService.pushNamedAndRemoveUntil(Routes.homeView);
+    await _routerService.router.push(
+      HomeRoute(),
+    );
   }
 }`;
   }